+++ 6CS002_Task_04/src/com/wlv/reflection/reflection_2.java
@@ -1,14 +1,17 @@
 package com.wlv.reflection;
 
+import java.lang.Double;
+
 public class reflection_2 {
   public static void main(String[] args) {
     Simple s = new Simple();
     s.rectangleA();
     
   // s.squareB();
-    double a = s.a;
+    float a = s.a;
   
   // double b = s.b; 
     System.out.println("s=" + s);
   }
 }
+
